requirements.txt
#general modules
django >3.2,<3.3
python-decouple==3.6
pillow
djangorestframework

#third-party modules
markdown       
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
djsoer
#deployment modules

# email third party modules
django-mail-templated

#refactoring and test
black
flake8